ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
Pillow==11.2.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
torchxrayvision
scikit-image
openai
matplotlib
seaborn
tqdm
pandas 